Add signal processing
sound_the_alarm.py now sends a signal to clock.py if the radio is to be played.
This way we can use the RadioStreamer class to keep track of the subprocess as well as set the _Play radio_ button as pressed. The signal handler also automatically
checks whether the displayed alarm time should be hidden for the weekend.